requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
ultralytics==8.3.0
opencv-python==4.8.1.78
Pillow==9.5.0
numpy==1.24.3
pydantic==2.4.2
torch==2.0.1